requirements.txt
anthropic
ccxt
pandas
pandas-ta
requests
python-dotenv
feedparserpsycopg2-binary
